+++ app/src/main/java/com/pns/bbaspassenger/view/dialogs/UserInfoDialog.kt
@@ -1,11 +1,10 @@
-package com.pns.bbaspassenger.view
+package com.pns.bbaspassenger.view.dialogs
 
 import android.app.Dialog
 import android.os.Bundle
 import android.util.Log
 import android.view.View
 import android.view.WindowManager
-import androidx.core.widget.doOnTextChanged
 import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.viewModels
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
