+++ app/src/main/java/com/pns/bbaspassenger/view/help/TutorialActivity.kt
@@ -1,4 +1,4 @@
-package com.pns.bbaspassenger.view
+package com.pns.bbaspassenger.view.help
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
